requirements.txt
pytest==6.2.3
PyYAML==5.4.1
selenium==3.141.0
zmail==0.2.8
allure-pytest==2.8.17
xlrd==1.2.0
openpyxl==3.0.7
Faker==8.4.0
pytest-rerunfailures==10.0
Appium-Python-Client~=1.1.0